packages/retrieval/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taboot-retrieval"
version = "0.4.0"
description = "Taboot LlamaIndex indices, retrievers, query engines"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Jacob Magar", email = "jmagar@gmail.com" }]
license = { text = "Proprietary" }

dependencies = [
  "llama-index-core>=0.14.4,<1",
  "llama-index-vector-stores-qdrant>=0.8.6,<1",
  "llama-index-graph-stores-neo4j>=0.5.1,<1",
  "llama-index-workflows>=2,<3",
  "llama-index-llms-ollama>=0.3.0,<1",
  "llama-index-embeddings-huggingface>=0.3.0,<1",
  "pydantic>=2.12.0,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["taboot_retrieval"]
